Properly format the timer

+++ app/src/main/java/com/illiouchine/toothbrush/ui/composable/chrono/Marker.kt
@@ -22,8 +22,8 @@ private const val NB_MARKER = 120
 
 @Composable
 internal fun CircleMarkers(
-    totalSeconds: Int,
-    seconds: Int,
+    totalSeconds: Long,
+    seconds: Long,
     markerColor: Color = MaterialTheme.colors.primaryVariant,
     activeMarkerColor: Color = MaterialTheme.colors.primary,
 ) {
